Allow binding to ::1 on loopback in any case
There is a strange bug / feature in getaddrinfo(..) that it fails
if the only interface that supports IPv6 is the loopback interface,
as this interface is being ignored when scanning for IPv6 support.

See also this RedHat issue:

https://bugzilla.redhat.com/show_bug.cgi?id=808147

This is a workaround so that binding to ::1 still works in this
scenario.

To test this, make sure that your system has all interfaces except
'lo' configured to be IPv4-only and run the following program:

(Code adapted from SF issue for libevent #361, see here:
 https://sourceforge.net/p/levent/bugs/361/)
---------------------------------------------------------------
 #include <event2/event.h>
 #include <event2/http.h>

int main()
{
    struct event_base* evb = event_base_new();
    struct evhttp* http = evhttp_new(evb);
    int res_ipv6 = evhttp_bind_socket(http, "::1", 5000);
    printf("%d\n", res_ipv6);
    return 0;
}
---------------------------------------------------------------

Result without this patch applied:
---------------------------------------------------------------
[warn] getaddrinfo: address family for nodename not supported
-1
---------------------------------------------------------------

Result with this patch applied:
---------------------------------------------------------------
0
---------------------------------------------------------------