Fix cache keys on CI
We use cache to make cmake runs faster, but this can break things, since
this will preserve all cmake cached variables, and it may not be updated
again.

And this is indeed what happened after the release, and this breaks
main/version test, first of all it reports 2.2.0, while it should be
2.2.1:

    --         ---( Libevent 2.2.0-alpha-dev )---

And then the test fails:

    FAIL /home/runner/work/libevent/libevent/dist/libevent-2.2.1-alpha-dev/test/regress.c:2632: assert((vint&0xffffff00) == ((major<<24)|(minor<<16)|(patch<<8))): 33685760 vs 33685504main/version: 1/432 TESTS FAILED. (42 skipped)

So for now I've added the tag name that assosiated with the commit, to
avoid this situation.

Also AFAIR it was significant to use cache for cmake, but next step will
be to remove it completelly..