Fix documentation inconsistency in evhttp callbacks
This commit addresses a minor inconsistency identified in the
documentation of multiple `evhttp` callback functions within
the `event2/http.h` header file. It was observed that the word
"additional" was presumably missing in the description of the
`arg` parameter for several functions. This oversight led to
an incorrect phrasing: "an context argument for the callback."

The documentation for the following functions has been updated
to rectify this issue:

  - `evhttp_set_gencb`
  - `evhttp_set_bevcb`
  - `evhttp_set_newreqcb`
  - `evhttp_set_errorcb`

This commit solely improves the readability of the function
descriptions without altering any functional aspects of the
code.