evdns: add DNS_CNAME_CALLBACK_FIRST option for A/AAAA CNAME callback
In common case A/AAAA callback with CNAME going after main answer, but
user can already process the first callback, free the request handler
and nothing known abount second callback.
To prevent the segmentation fault in CNAME callback we can revert order
of callbacks with this option and control behaviour by type of request
in callback (A/AAAA or CNAME).