evdns: add CNAME type resolver and DNS_CNAME_CALLBACK_FIRST flag
- Add CNAME type call for resolver.
- Add DNS_CNAME_CALLBACK_FIRST flag for A/AAAA CNAME callback.

In common case A/AAAA callback with CNAME going after main answer, but
user can already process the first callback, free the request handler
and nothing known abount second callback.
To prevent the segmentation fault in CNAME callback we can revert order
of callbacks with this flag and control behaviour by type of request
in callback (A/AAAA or CNAME).